Refactored navbar and routing
